\name{XC}

\alias{XC}

  
\title{Compute the XC score}

\description{
Compute the XC score
}

\usage{
XC(regions, sample, env=NULL, envfun=NULL, minssize=10, maxdist=1500, return_network=FALSE)
}

\arguments{
\item{regions}{SpatVector or polygons}
\item{sample}{SpatVector or points}
\item{env}{SpatRaster with environmental data or NULL}
\item{envfun}{function}
\item{minssize}{numeric}
\item{maxdist}{numeric}
\item{return_network}{logical}
}

\value{
numeric 
}

\examples{
2 + 2
}

\keyword{spatial}

